requirements.txt
Django>=4.2.0,<4.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2==2.9.9
drf-spectacular>=0.26.5,<0.27
Pillow>=10.1.0,<10.2.0
django-filter==23.4
django-debug-toolbar==4.2.0
